1. Changed the IP address to my IP address, put yours in comment
2. Removed the Scanner for email and password and got the email and password from the
database using the Database_accessor.getiMessageUser(username) method.
The username passed through that method is obtained through the iMessageUser object
created in Main, that can be accessed and updated in any class using Main.currentUser.getUsername().
The database is accessed to get the password for this account to pass into the client.signIn parameters.

3. I am going to remove the scanner for recipient and make that a clickable method from
the chat_screen_controller.

4. messageContent scanner also needs to be removed and implemented through the textfield in the
chat_screen_controller.
client

5. Number 3 and 4 are done by removing these scanner lines and creating a method in
chat_screen_controller that returns the recipient, based on who is selected from the
tableview, and returns messageContent, based on what is obtained from the textfield.